Parallelize test generation
